requirements.txt
jsonschema==3.0.0a3
pytest==4.3.0
PyYAML==3.13
